pyproject.toml
[project]
name = "bus-seating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
